==> scripts/50_pt_macroecologie.Rmd <==
---
title: "Macroécologie au point"
author: "OFB"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
output:
  bookdown::html_document2:
    toc: true
    toc_float: true
#  bookdown::word_document2
always_allow_html: true
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, message=F, warning = F)
```

# Chargement des packages et des données

```{r}
library(tidyverse)
library(sf)
library(vegan)
library(corrplot)
library(factoextra)
library(FactoMineR)
library(mapview)
library(aspe)
library(COGiter)

source(file = "../R/lm_indices_env.R")
source(file = "../R/mon_accum.R")

```


```{r}
load(file = "../processed_data/pt_carto_data.rda")
load(file = "../processed_data/aspe_macroecologie.rda")
```

# Indices de biodiversité

Dans cette partie, seules les données d'inventaires de la base Aspe sont conservées. Pour les points prospectés plusieurs fois, on ne conserve que la dernière pêche en date pour éviter un biais d'effort de prospection.

## Distribution des indices

NB Pielou n'est pas calculé sur les points monospécifiques car il a au dénominateur le log de la richesse.

```{r}
pop_indices_long %>% 
  mutate(name = fct_relevel(name, "richesse", "shannon", "simpson")) %>% 
  ggplot(aes(x = value)) + 
  geom_histogram(fill = "darkgreen",
                 col = "black",
                 alpha = 0.2,
                 bins = 20) +
  scale_x_continuous(limits = c(0, NA)) +
  facet_wrap(~name, scales = "free_x") +
  theme_classic() +
  labs(x = "Valeur de l'indice",
       y = "Nombre de sites")
```


## Cartographie des indices

```{r, fig.width = 5, fig.height = 12}
test <- aspe_pops_geo %>%
 # rename(code_point = pop_id) %>% 
  right_join(pop_indices_long) #%>% 
            #   mutate(code_point = as.integer(code_point)))

pal <- wesanderson::wes_palette("Zissou1", 100, type = "continuous")

bg <- COGiter::departements_metro_geo %>% 
  filter(DEP %in% c("22", "29", "35", "56"))

g <- purrr::map(.x = unique(test$name),
                .f = function(x) {
                            ggplot() +
                    geom_sf(data = bg) +
                    geom_sf(data = filter(test, name == x),
                                   aes(col = value)) +
                    ggtitle(x) +
                    scale_color_gradientn(colours = pal,
                                          name = "")
                })

g2 <- cowplot::plot_grid(plotlist = g,
                         ncol = 1)

g2
```

# Variables environnementales

## Examen des variables explicatives

```{r}
library("PerformanceAnalytics")
chart.Correlation(pop_indices %>% select(altitude:y_wgs84),
                  histogram = TRUE,
                  pch = 19)
```

Pour se rapprocher de distributions normales, on va log-transformer les variables à part les coordonnées.

```{r}
pop_indices <- pop_indices %>%
  mutate_at(vars(dist_source:pente, starts_with("temp")), log) %>% 
  mutate(altitude = log(1 + altitude))

chart.Correlation(pop_indices %>% select(altitude:y_wgs84),
                  histogram = TRUE,
                  pch = 19)
```

Les distributions sont beaucoup plus satisfaisantes. Les variables `dist_source`, `surf_bv` et `largeur` sont trop corrélées pour être utilisées telles quelles dans les modèles donc on va les regrouper en une variable synthétique par ACP.

## Construction de l'indice de taille du cours d'eau

```{r}
taille <- FactoMineR::PCA(pop_indices %>% select(dist_source, surf_bv, largeur))
taille <- taille$ind$coord[,1]
```

Ajout de la variable `taille` au dataframe.

```{r}
pop_indices <- pop_indices %>% 
  cbind(taille)
```

Vérification.

```{r}
cor(pop_indices %>% select(taille, dist_source, surf_bv, largeur))
```

# Modélisation

## Richesse spécifique

On construit un modèle avec l'ensemble des variables explicatives, puis on le simplifie par une procédure pas à pas au moyen de la fonction `stepAIC()` du package `MASS`.

### Modèle complet

```{r}
mod <- lm(richesse ~ altitude +
            taille +
            pente +
            temp_01 +
            temp_07 +
            x_wgs84 +
            y_wgs84,
          data = pop_indices)

summary(mod)
```

### Simplification

```{r, results='hide'}
mod2 <- MASS::stepAIC(mod)
```


```{r}
summary(mod2)
```

### Interprétation

On observe donc que les stations sont d'autant plus riches qu'elles sont :

- sur des cours d'eau grands, peu pentus, frais en hiver et chauds en été
- localisées au sud de l'aire étudiée

Bien que le paramètre de longidude `x_wgs84` ne soit pas retenu par la procédure de simplification, les deux paramètres de température correspondent à un gradient de longitude. En effet sur l'Ouest de la zone d'étude les hivers sont plus doux et les étés plus frais qu'à l'Est.

#### Vérifications graphiques

```{r}
plot(mod2)
```

## Autres indices

Sans aller autant dans le détail que pour la richesse spécifique on peut reproduire la démarche pour les autres indices (sans appliquer ici la `stepAIC`; on conserve toutes les variables explicatives).

```{r}
lm_indices_env(
  df = pop_indices,
  variables_dependantes = c("richesse",
                            "shannon",
                            "simpson",
                            "pielou"),
  variables_explicatives = c("taille",
                             "pente",
                             "temp_01",
                             "temp_07",
                             "x_wgs84",
                             "y_wgs84"),
  step = FALSE # ne peut être TRUE que si tous les modèles simplifiés par stepAIC retiennent les mêmes variables
) %>%
  rownames_to_column("Variable explicative") %>%
  flextable::flextable() %>%
  flextable::set_table_properties(layout = "autofit", width = .9) %>%
  flextable::theme_zebra() %>% 
  flextable::hline(i = c(1, 7)) %>% 
  flextable::italic(i = 8) %>% 
  flextable::bold(i = 8)
```

# Problématique de l'échantillonnage

## Estimation de la richesse absolue par point

On utilise l'exemple fourni sur 
https://search.r-project.org/CRAN/refmans/vegan/html/specpool.html

On reprend toutes les opérations pour chaque point.

```{r}
ope_abondance_large <- ope_effectif %>% 
  pivot_wider(names_from = "code_espece",
              values_from = "effectif",
              values_fill = 0) %>% 
  left_join(aspe_passerelle %>% 
              select(ope_id,
                     pop_id) %>% 
              unique())

# parenthèse en vue de l'analyse par BV
bv_abondance_large <- ope_abondance_large %>% 
  filter(ope_id %in% ope_id_der_peche_par_pop_id) %>% 
  left_join(y = pop_indices %>% 
                  select(pop_id, code_exutoire) %>% 
                  unique()) %>% 
  left_join(y = aspe_pops_geo %>%
              sf::st_drop_geometry() %>% 
              select(code_exutoire,
                     toponyme))

# mise en forme
ope_abondance_large <- ope_abondance_large %>% 
    select(-ope_id, -pro_libelle)

# calcul de la richesse et mise au propre
pop_richesse_absolue <- specpool(x = ope_abondance_large %>% select(-pop_id),
                                 pool = ope_abondance_large$pop_id)

pop_richesse_absolue <- pop_richesse_absolue %>% 
  rownames_to_column("pop_id") %>% 
  mutate(pop_id = as.integer(pop_id)) %>% 
    select(-ends_with(".se")) %>% 
  rename(richesse_observee = Species,
         n_ope = n) %>% 
 #   rownames_to_column(var = "site") %>% 
    mutate(richesse_absolue_moy_estimateurs = (chao + jack1 + jack2 + boot) / 4) %>% 
    mutate_all(round, 2) %>% 
    left_join(y = pt_carto_data %>% 
                  select(pop_id = code_point,
                         localisation,
                         code_exutoire,
                         bassin = toponyme
                         ) %>%
                  distinct()) %>% 
  select(pop_id,
         localisation,
         code_exutoire,
         bassin,
         n_ope,
         richesse_observee,
         richesse_absolue_moy_estimateurs,
         everything())

DT::datatable(pop_richesse_absolue)
```

## Courbes d'accumulation / appauvrissement

Calculées pour les sites avec mini 30 pêches

```{r, fig.height = 8, fig.width = 8}
accum <- mon_accum (df = ope_abondance_large,
                    var_site_id = pop_id)

points_assez_de_peches <- ope_abondance_large %>% 
  count(pop_id) %>% 
  filter(n > 30) %>% 
  pull(pop_id)


ggplot(data = accum %>% filter(pop_id %in% points_assez_de_peches),
       aes(x = n_opes,
           y = pt_richesse)) +
  geom_line() +
  facet_wrap(~pop_id) +
  labs(x = "Nombre de pêches")
```


```{r, fig.height = 8, fig.width = 8}
accum %>%
  filter(pop_id %in% points_assez_de_peches) %>% 
  left_join(y = pt_carto_data %>% 
              select(pop_id = code_point,
                     localisation) %>% 
              distinct()) %>% 
              mutate(localisation = as.character(localisation),
                     localisation = ifelse(is.na(localisation), "Inconnu", localisation),
                     localisation = paste0(pop_id, " (", localisation, ")")) %>%
ggplot(aes(x = n_opes,
           y = pt_richesse,
           col = localisation)) +
  geom_line() +
  labs(x = "Nombre de pêches",
       col = "") +
  theme(legend.position = "bottom")
```


# Sauvegarde des résultats

```{r}
save(bv_abondance_large,
     file = "../processed_data/bv_abondance_large.rda")
save(pop_richesse_absolue,
     file = "../processed_data/pop_richesse_absolue.rda")
```
