==> scripts/51_bv_macroecologie.Rmd <==
---
title: "Macroécologie bassins"
author: "OFB"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
output:
  bookdown::html_document2:
    toc: true
    toc_float: true
#  bookdown::word_document2
always_allow_html: true
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, message=F, warning = F)
```

# Chargement des packages et des données

```{r}
#library(MASS) # keep MASS first yo avoid synonymous function conflict with dplyr
library(tidyverse)
#library(sf)
library(vegan)
library(corrplot)
library(factoextra)
library(FactoMineR)
library(mapview)
#library(aspe)


source(file = "../R/lm_indices_env.R")
```


```{r}
#load(file = "../processed_data/data2.rda")
# load(file = '../processed_data/aspe.RData')
# load(file = "../processed_data/especes_a_supprimer.rda")
load(file = "../processed_data/carto_data2.rda")
load(file = "../processed_data/bv_abondance_large.rda")
```

# Cartographie des richesses par bassin

```{r}
bv_richesse <- bv_faune %>% 
  filter(statut == "Présent") %>% 
  group_by(code_exutoire) %>% 
  summarise(rsr = n_distinct(code_espece))

map_data <- bv_simp_geo %>% 
  right_join(bv_richesse) %>% 
  dplyr::select(code_exutoire,
         toponyme,
         rsr)

mapview(map_data,
        zcol = "rsr")
```

# Explications de la richesse des bassins

## Assemblage des données

```{r}
bv_data <- bv_env %>% 
    mutate(surf_km2 = surf_m2 / 1e6,
           long_tp_km = long_tp_m / 1000) %>% 
    mutate_at(vars(starts_with("X_")), as.numeric) %>% 
  select(-surf_m2,
         -long_tp_m) %>% 
  left_join(bv_effort_prospection) %>% 
  left_join(bv_richesse) %>% 
  filter(rsr > 0,
         n_inventaires > 0)
```

## Distribution des variables explicatives

Distribution des variables quantitatives. Les abscisses sont en échelle log.

```{r, fig.width = 8, fig.height = 8}
plot_data <- bv_data %>% 
  select(code_exutoire,
         alt_moy:surf_km2,
         n_inventaires,
         n_pops) %>% 
  pivot_longer(cols = -code_exutoire) %>% 
  mutate(name = as.factor(name),
         name = fct_relevel(name, "n_inventaires","n_pops"))


ggplot(data = plot_data %>% filter(!name %in% c("n_inventaires","n_pops")),
         aes(x = value#,
             #col = name %in% c("n_inventaires","n_pops")
             )) +
    geom_histogram(fill = "darkgreen") +
    facet_wrap(~name, scales = "free") +
    scale_x_log10(labels = scales::label_comma()) +
    theme(legend.position="none")
```

Pour ce qui est de l'effort de prospection, la Vilaine est très loin des autres bassins, ca qui "tasse" les distributions des petites valeurs. 

```{r, fig.height = 3, fig.width = 4}
ggplot(data = plot_data %>% filter(name %in% c("n_inventaires","n_pops")),
         aes(x = value#,
             #col = name %in% c("n_inventaires","n_pops")
             )) +
    geom_histogram(fill = "darkred") +
    facet_wrap(~name, scales = "free") +
    theme(legend.position="none")
```


Variables qualitatives

```{r}
bv_data %>% 
  select(code_exutoire,
         canal_conn,
         strahler_m) %>% 
  pivot_longer(cols = -code_exutoire) %>% 
  ggplot(aes(x = value)) +
    geom_bar(fill = "darkgreen") +
    facet_wrap(~name, scales = "free")
```

## Relations entre variables explicatives

On se limite aux variables quantitatives car d'une part la connexion à un canal est très rare et d'autre part le rang de strahler ne sera pas conservé car son information serait redondante avec celle des variables quantitatives.

```{r}
library("PerformanceAnalytics")
chart.Correlation(bv_data %>% select(alt_moy:surf_km2, n_inventaires, n_pops),
                  histogram = TRUE,
                  pch = 19)
```

Test en log-transformant.

```{r}
chart.Correlation(bv_data %>%
                    select(alt_moy:surf_km2, n_inventaires, n_pops) %>%
                    mutate_at(vars(alt_moy:surf_km2, n_inventaires, n_pops), log10),
                  histogram = TRUE,
                  pch = 19)
```

Les distributons sont plus proches de la normale quand les variables "surf_km2", "prct_PE_CE" et "pte_tp_moy" sont log-transformées, donc on applique cette transformation.

```{r}
bv_data <- bv_data %>% 
  mutate(prct_PE_CE = ifelse(is.na(prct_PE_CE), 0, prct_PE_CE),
         prct_PE_CE = log(1 + prct_PE_CE)) %>% 
  mutate_at(vars(surf_km2, pte_tp_moy, n_inventaires, n_pops), log)
```

```{r}
acp <- bv_data %>% 
  select(code_exutoire,
         alt_moy:surf_km2) %>% 
  column_to_rownames("code_exutoire") %>% 
  FactoMineR::PCA(graph = FALSE)

factoextra::fviz_pca_var(acp, col.var = "contrib", gradient.col = c("pink", "purple", "blue"), repel = TRUE)
```

```{r, fig.width = 8, fig.height = 8}
matrice_cor <- cor(bv_data %>% 
  select(alt_moy:surf_km2, n_pops, n_inventaires) %>% 
    na.omit())

corrplot(matrice_cor,
         addCoef.col = 1,
         number.cex = 0.6)
```

*A affiner* : on retient pour :

- l'altitude : la médiane, très corrélée à la moyenne mais moins corrélée à surf_km2
- la pente : la médiane, la moins corrélée à alt_max
- parc : la moyenne, très corrélée à la médiane et à l'écart-type donc qui synthétise cet aspect
- les variables prct_PE_CE, prct_rpg et surf_km2 qui sont peu corrélées aux autres
- on écarte la variable de connectivité avec des canaux, qui concerne trop peu de bassins (4) 

Ces variables environnementales sont complétées par la longitude et le nombre de points d'échantillonnage dans le bassin.

## Modélisation

Une série de tests préalables a conduit à écarter certains bassins - les plus grands - qui n'entrent pas dans le schéma général.

```{r}
model_data <- bv_data %>% 
  filter(n_inventaires > 0)

rownames(model_data) <- paste(model_data$toponyme, model_data$code_exutoire, sep = " -")
```

### Modèle sur tous les bassins

```{r}
les_variables_explicatives = c(
  "alt_median",
  "pente_medi",
  "parc_moy",
  "prct_PE_CE",
  "prct_rpg",
  "surf_km2",
  "X_centroid",
  "log(n_inventaires)",
  "canal_conn"
)
# 
mod0 <- lm(data = model_data,
           formula = paste("rsr ~ ",
                           paste(les_variables_explicatives, collapse = " + ")))

summary(mod0)
```

Simplification

```{r, results='hide'}
mod0step <- MASS::stepAIC(mod0)
```


```{r}
summary(mod0step)
```

```{r}
plot(mod0step)
```

### Sans la Vilaine

On voit que la Vilaine est un outlier qui influence très fortement le modèle (fort bras de levier `leverage`). Elle n'entre donc pas vraiment dans le schéma général.

Observons ce qui se passe si l'on supprime la Vilaine.

```{r, results='hide'}
mod1 <- lm(data = model_data %>% 
             filter(!str_detect(toponyme, "Vilaine")) ,
           formula = paste("rsr ~ ",
                           paste(les_variables_explicatives, collapse = " + ")))
mod1step <- MASS::stepAIC(mod1)
```


```{r}
summary(mod1step)

plot(mod1step)
```

### En supprimant Le Cousenon, la Sélune et la rivière du Bono

```{r, results='hide'}
mod2 <- lm(data = model_data %>% 
             filter(!str_detect(toponyme, "Vilaine|du Bono|Couesnon|La Sã©Lune")) ,
           formula = paste("rsr ~ ",
                           paste(les_variables_explicatives, collapse = " + ")))
mod2step <- MASS::stepAIC(mod2)
```


```{r}
summary(mod2step)

plot(mod2step)
```

### En supprimant la Blavet et l'Aulne

```{r, results='hide'}
mod3 <- lm(data = model_data %>% 
             filter(!str_detect(toponyme, "Vilaine|du Bono|Couesnon|La Sã©Lune|Blavet|Aulne")) ,
           formula = paste("rsr ~ ",
                           paste(les_variables_explicatives, collapse = " + ")))
mod3step <- MASS::stepAIC(mod3)
```


```{r}
summary(mod3step)

plot(mod3step)
```

### Comparaison des modèles

Qu'on considère le modèle construit sur l'ensemble du jeu de données ou bien ceux omettant les bassins les plus influents, les variables retenues sont toujours les mêmes, ce qui est rassurant. On va comparer ces modèles pour voir dans quelle mesure l'omission de certains bassins influe sur les modèles. On ne conserve que les 4 variables qui sont sysnématiquement retenues par la procédure de sélection `stepAIC`.

```{r}
les_variables_explicatives = c(
  "alt_median",
 # "pente_medi",
 # "parc_moy",
#  "prct_PE_CE",
  #"prct_rpg",
  "surf_km2",
  "X_centroid",
  "log(n_inventaires)"#,
  #"canal_conn"
)
test <- map(.x = list(model_data,
                      model_data %>% filter(!str_detect(toponyme, "Vilaine")),
                      model_data %>% filter(!str_detect(toponyme, "Vilaine|du Bono|Couesnon|La Sã©Lune")),
                      model_data %>% filter(!str_detect(toponyme, "Vilaine|du Bono|Couesnon|La Sã©Lune|Blavet|Aulne"))
                    ),
            .f = lm_indices_env,
            variables_dependantes = "rsr",
            variables_explicatives = les_variables_explicatives,
            step = FALSE) %>% 
  reduce(cbind) %>% 
  as.data.frame() %>% 
  set_names(c("Modèle complet", "Sans Vilaine", "Exclusion 4 BV", "Exclusion 6 BV")) %>% 
  rownames_to_column(var = "Variable explicative")

flextable::flextable(test) %>% 
  flextable::set_table_properties(layout = "autofit", width = .9) %>% 
  flextable::bg(i = 6, bg = "#EFEFEF", part = "body")
```

Il apparaît donc que, bien que certains grands bassins contribuent fortement à construire les modèles, leur inclusion ou exclusion du jeu de données ne change que très peu la valeur des coefficients. Aucun de change de signe. Il n'est donc pas choquant de conserver le modèle avec tous les bassins même la Vilaine.

### Interprétation

Le coefficient de détermination ajusté du modèle est de `r test[2, 6]`. La richesse d'un bassin est d'autant plus élevé qu'il est :

- grand
- bien prospecté
- localisé à l'Est de la zone d'étude
- à faible attitude

# Problématique de l'échantillonnage

## Estimation de la richesse absolue par bassin

On utilise l'exemple fourni sur 
https://search.r-project.org/CRAN/refmans/vegan/html/specpool.html

```{r}
bv_richesse_absolue <- specpool(x = bv_abondance_large %>% select(-pop_id,
                                                                  -ope_id,
                                                                  -pro_libelle,
                                                                  -code_exutoire,
                                                                  -toponyme),
                                pool = bv_abondance_large$toponyme)
```

Un peu de nettoyage / renommage.

```{r}
bv_richesse_absolue <- bv_richesse_absolue %>% 
  rownames_to_column("toponyme") %>% 
    select(-ends_with(".se")) %>% 
  rename(richesse_observee = Species,
         n_pop = n) %>% 
  filter(!is.na(richesse_observee))
```

Suppression des valeurs aberrantes des indicateurs de richesse absolue (inférieures à la richesse observée - cas très rare). 

```{r}
bv_richesse_absolue <- bv_richesse_absolue %>% 
  mutate_at(vars(chao:boot),
            function(x) ifelse(x < bv_richesse_absolue$richesse_observee, NA, x))
```

Calcul d'un indicateur "moyen" `richesse_absolue_moy_estimateurs`.

```{r}
bv_richesse_absolue <- bv_richesse_absolue %>% 
  mutate(richesse_absolue_moy_estimateurs = rowMeans(bv_richesse_absolue %>% 
                                                       select(chao:boot),
                                                     na.rm = T)) %>% 
  mutate_if(is.numeric, round, 2) %>% 
  select(code_exutoire,
         n_pop,
         rich_obs = richesse_observee,
         rich_abs_moy = richesse_absolue_moy_estimateurs,
         everything())

DT::datatable(bv_richesse_absolue)

GGally::ggpairs(bv_richesse_absolue %>% 
                  select(-code_exutoire, -n_pop))

ggplot(data = bv_richesse_absolue,
       aes(x = rich_obs,
           y = rich_abs_moy)) +
  geom_point() +
  #geom_smooth(method = "lm") +
  geom_abline(slope = 1, intercept = 0, col = "red")
```

# Sauvegarde des résultats

```{r}
save(bv_richesse_absolue,
     file = "../processed_data/bv_richesse_absolue.rda")
```
