==> scripts/52_lsr_rsr.Rmd <==
---
title: "Relation richesse locale - richesse régionale"
author: "OFB"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
output:
  bookdown::html_document2:
    toc: true
    toc_float: true
#  bookdown::word_document2
always_allow_html: true
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, message=F, warning = F)
```

# Chargement des packages et des données

```{r}
library(tidyverse)
```


```{r}
load(file = "../processed_data/pop_richesse_absolue.rda")
load(file = "../processed_data/bv_richesse_absolue.rda")
```

# Relation richesse locale - richesse BV

```{r}
bv_mean_lsr <- pop_richesse_absolue %>% 
  group_by(code_exutoire) %>% 
  summarise(richesse_moy = mean(richesse_absolue_moy_estimateurs))

loc_reg <- bv_richesse_absolue %>% 
  select(code_exutoire,
         n_pop,
         rsr_obs = rich_obs,
         rsr_abs = rich_abs_moy) %>% 
  left_join(bv_mean_lsr) %>% 
  rename(lsr_moy = richesse_moy)
```



```{r}
test <- ggplot(data = loc_reg,
       aes(x = rsr_abs,
           y = lsr_moy)) +
  geom_point() +
  stat_smooth(method="lm", se=FALSE, fill=NA,
                formula=y ~ poly(x, 2, raw=TRUE), colour="red") +
  stat_smooth(method = "nls",
              formula = y ~ SSasymp(x, Asym, R0, lrc), se = FALSE) +
  scale_y_continuous(limits = c(0, NA)) +
  geom_abline(intercept = 0, slope = 1, col = "darkgreen")

plotly::ggplotly(test)
```
